requirements.txt
fastapi
uvicorn
pychromecast
gtts
python-slugify
pydub